+++ src/liberty/projectTreeProvider.ts
@@ -1,0 +1,299 @@
+/*
+ * IBM Confidential
+ * Copyright IBM Corp. 2020, 2026
+ */
+import * as fse from "fs-extra";
+import * as vscode from "vscode";
+import * as util from "../util/helperUtil";
+import { localize } from "../util/i18nUtil";
+import { devModeRequirement } from "../util/helperUtil";
+import { UNTITLED_WORKSPACE } from "../definitions/constants";
+import { LibertyProject } from "./libertyProject";
+import { ProjectRegistry } from "./projectRegistry";
+import { discoverWorkspace } from "./projectDiscovery";
+
+// URI scheme used to signal dev-mode state to the FileDecorationProvider (enables green UI)
+// Format: liberty-dev://running/<encoded-project-path>
+const LIBERTY_DEV_SCHEME = "liberty-dev";
+
+export class LibertyDevDecorationProvider {
+	private _onDidChangeFileDecorations = new vscode.EventEmitter<vscode.Uri | vscode.Uri[]>();
+	readonly onDidChangeFileDecorations = this._onDidChangeFileDecorations.event;
+
+	public notify(uri: vscode.Uri): void {
+		this._onDidChangeFileDecorations.fire(uri);
+	}
+
+	provideFileDecoration(uri: vscode.Uri): { color: vscode.ThemeColor; tooltip: string } | undefined {
+		if (uri.scheme === LIBERTY_DEV_SCHEME && uri.authority === "running") {
+			return {
+				color: new vscode.ThemeColor("debugIcon.startForeground"),
+				tooltip: "Dev mode is running",
+			};
+		}
+		return undefined;
+	}
+}
+
+export class ProjectTreeProvider implements vscode.TreeDataProvider<LibertyProject> {
+	public readonly onDidChangeTreeData: vscode.Event<LibertyProject | undefined>;
+
+	private static instance: ProjectTreeProvider;
+
+	private _onDidChangeTreeData: vscode.EventEmitter<LibertyProject | undefined>;
+
+	private _refreshing = false;
+
+	private _treeView: vscode.TreeView<LibertyProject> | undefined;
+
+	private _registry: ProjectRegistry;
+
+	public readonly decorationProvider = new LibertyDevDecorationProvider();
+
+	constructor(registry: ProjectRegistry) {
+		this._registry = registry;
+		this._onDidChangeTreeData = new vscode.EventEmitter<LibertyProject | undefined>();
+		this.onDidChangeTreeData = this._onDidChangeTreeData.event;
+		this.refresh();
+	}
+
+	public static getInstance(): ProjectTreeProvider {
+		return ProjectTreeProvider.instance;
+	}
+
+	public static setInstance(provider: ProjectTreeProvider): void {
+		ProjectTreeProvider.instance = provider;
+	}
+
+	public async refresh(): Promise<void> {
+		if (this._refreshing) { return; }
+		this._refreshing = true;
+		this.setLoading(true);
+		const statusMessage = vscode.window.setStatusBarMessage(localize("refreshing.liberty.dashboard"));
+		const t0 = Date.now();
+		try {
+			await this.updateProjects();
+			console.log(`[perf] refresh total: ${Date.now() - t0}ms`);
+			this._onDidChangeTreeData.fire(undefined);
+		} finally {
+			statusMessage.dispose();
+			this.setLoading(false);
+			this.setMessage("");
+			this._refreshing = false;
+		}
+	}
+
+	/**
+	 * Update the tree item to reflect the new dev mode state.
+	 */
+	public notifyDevModeChanged(project: LibertyProject): void {
+		if (project.isDevMode) {
+			project.description = localize("liberty.view.running");
+			project.resourceUri = vscode.Uri.parse(
+				`${LIBERTY_DEV_SCHEME}://running/${encodeURIComponent(project.path)}`
+			);
+		} else {
+			project.description = undefined;
+			project.resourceUri = undefined;
+		}
+		this._onDidChangeTreeData.fire(project);
+		if (project.resourceUri) {
+			this.decorationProvider.notify(project.resourceUri);
+		}
+	}
+
+	public fireChangeEvent(): void {
+		this._onDidChangeTreeData.fire(undefined);
+	}
+
+	public setTreeView(treeView: vscode.TreeView<LibertyProject>): void {
+		this._treeView = treeView;
+	}
+
+	public setMessage(message: string): void {
+		if (this._treeView) {
+			this._treeView.message = message;
+		}
+	}
+
+	private setLoading(loading: boolean): void {
+		vscode.commands.executeCommand('setContext', 'liberty:loading', loading);
+	}
+
+	/**
+	 * This method asks the user to save the workspace first if it is untitled and contains
+	 * more than one project.
+	 */
+	public async checkUntitledWorkspaceAndSaveIt(): Promise<void> {
+		return new Promise((resolve) => {
+			try {
+				vscode.window.showInformationMessage(
+					localize("workspace.not.saved.projects.may.not.persist"),
+					{ modal: true },
+					'Save Workspace'
+				).then(async (selection) => {
+					if (selection === 'Save Workspace') {
+						await this._registry.getContext().globalState.update('workspaceSaveInProgress', true);
+						await vscode.commands.executeCommand('workbench.action.saveWorkspaceAs');
+					}
+					util.clearDataSavedInGlobalState(this._registry.getContext());
+					resolve();
+				});
+			} catch (error) {
+				console.debug("exception while saving the workspace" + error);
+				util.clearDataSavedInGlobalState(this._registry.getContext());
+				resolve();
+			}
+		});
+	}
+
+	public isMultiProjectUntitledWorkspace(): boolean {
+		const workspaceFolders = vscode.workspace.workspaceFolders;
+		if ((workspaceFolders && workspaceFolders.length > 1
+			&& vscode.workspace.name === UNTITLED_WORKSPACE)) {
+			return true;
+		}
+		return false;
+	}
+
+	public getTreeItem(element: LibertyProject): vscode.TreeItem {
+		element.collapsibleState = (element.isAggregator && element.children.length > 0)
+			? vscode.TreeItemCollapsibleState.Expanded
+			: vscode.TreeItemCollapsibleState.None;
+		return element;
+	}
+
+	// eslint-disable-next-line @typescript-eslint/no-unused-vars
+	public async getChildren(element?: LibertyProject): Promise<LibertyProject[]> {
+		if (vscode.workspace.workspaceFolders === undefined) {
+			vscode.window.showInformationMessage(localize("no.liberty.project.found.in.empty.workspace"));
+			return [];
+		}
+		if (element === undefined) {
+			const rootProjects = this._registry.getRootProjects();
+			if (rootProjects.length > 0) {
+				return rootProjects;
+			}
+			return [...this._registry.getProjects().values()];
+		}
+		return element.children.filter(child =>
+			child.isLibertyEnabled || this.hasLibertyDescendants(child)
+		);
+	}
+
+	/**
+	 * Unified project picker. Handles command palette, aggregator delegation, and direct selection.
+	 */
+	public async pickProject(project: LibertyProject | undefined, command: string): Promise<LibertyProject | undefined> {
+		const placeholder = localize("select.module.for.command");
+
+		if (project === undefined) {
+			const { filterProjects } = await import("../util/helperUtil");
+			const projects = filterProjects(Array.from(this._registry.getProjects().values()), command);
+			if (projects.length === 0) {
+				const message = localize("no.liberty.projects.found");
+				console.error(message);
+				vscode.window.showInformationMessage(message);
+				return undefined;
+			}
+			if (projects.length === 1) {
+				return projects[0];
+			}
+			const items = projects.map(p => ({
+				label: p.label,
+				description: p.parent ? p.parent.label : undefined,
+				detail: p.path,
+				project: p,
+			}));
+			const selected = await vscode.window.showQuickPick(items, { placeHolder: placeholder });
+			return selected?.project;
+		}
+
+		if (project.isAggregator && project.children.length > 0) {
+			const libertyChildren = this._registry.findLibertyDescendants(project);
+			if (libertyChildren.length === 0) {
+				vscode.window.showWarningMessage(
+					localize("no.liberty.modules.found", project.label)
+				);
+				return undefined;
+			}
+			const req = devModeRequirement(command);
+			const eligible = req === undefined
+				? libertyChildren
+				: libertyChildren.filter(c => c.isDevMode === req);
+			if (eligible.length === 0) {
+				const msg = req === true
+					? localize("no.modules.currently.running", project.label)
+					: localize("all.modules.currently.running", project.label);
+				vscode.window.showInformationMessage(msg);
+				return undefined;
+			}
+			if (eligible.length === 1) {
+				return eligible[0];
+			}
+			const items = eligible.map(c => ({
+				label: c.label,
+				description: c.parent ? c.parent.label : undefined,
+				detail: c.path,
+				project: c,
+			}));
+			const selected = await vscode.window.showQuickPick(items, { placeHolder: placeholder });
+			return selected?.project;
+		}
+
+		if (project.isLibertyEnabled) {
+			return project;
+		}
+
+		vscode.window.showWarningMessage(
+			localize("project.not.liberty.enabled", project.label)
+		);
+		return undefined;
+	}
+
+	private hasLibertyDescendants(project: LibertyProject): boolean {
+		if (project.isLibertyEnabled) { return true; }
+		return project.children.some(child => this.hasLibertyDescendants(child));
+	}
+
+	private async updateProjects(): Promise<void> {
+		const t0 = Date.now();
+		const context = this._registry.getContext();
+		const wsFolders = vscode.workspace.workspaceFolders ?? [];
+
+		// Clean up persisted projects whose files no longer exist
+		const persistedProjects = util.getStorageData(context).projects;
+		const persistedToRemove: typeof persistedProjects[0][] = [];
+		for (const p of persistedProjects) {
+			if (!fse.existsSync(p.path)) {
+				persistedToRemove.push(p);
+			}
+		}
+		if (persistedToRemove.length > 0) {
+			const dashboardData = util.getStorageData(context);
+			persistedToRemove.forEach(p => dashboardData.removeProject(p.path));
+			await util.saveStorageData(context, dashboardData);
+		}
+		const validPersisted = persistedProjects.filter(p => !persistedToRemove.includes(p));
+
+		const { projects, rootProjects } = await discoverWorkspace(
+			context,
+			wsFolders,
+			this._registry.getProjects(),
+			validPersisted,
+			(partial, foldersComplete, totalFolders) => {
+				// Progressive tree update after each folder
+				this._registry.setProjects(new Map(partial));
+				this._registry.setRootProjects(Array.from(partial.values()).filter(p => !p.parent));
+				if (foldersComplete < totalFolders) {
+					this.setMessage(localize("scanning.workspace.folders", foldersComplete, totalFolders));
+				}
+				this._onDidChangeTreeData.fire(undefined);
+			}
+		);
+
+		this._registry.setProjects(projects);
+		this._registry.setRootProjects(rootProjects);
+		console.log(`[perf] updateProjects total: ${Date.now() - t0}ms  (${projects.size} projects, ${rootProjects.length} roots)`);
+	}
+}
