+++ src/locales/en.json
@@ -10,6 +10,7 @@
     "jakarta.ls.output.dropdown": "Language Support for Jakarta EE",
     "jakarta.ls.thumbs.up": "LSP4Jakarta $(thumbsup)",
     "jakarta.ls.started": "Language Server for Jakarta EE started.",
+    "liberty.view.running": "Running...",
     "no.liberty.projects.found": "No Liberty projects found.",
     "starting.liberty.dev.on": "Starting Liberty dev mode on {0}.",
     "cannot.start.liberty.dev": "Cannot start Liberty dev mode on an undefined project.",
